# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtistDetailOut": {
        "properties": {
          "artist": {
            "$ref": "#/components/schemas/ArtistOut"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/ContractOut"
            },
            "title": "Contracts",
            "type": "array"
          },
          "releases": {
            "items": {
              "$ref": "#/components/schemas/ReleaseOut"
            },
            "title": "Releases",
            "type": "array"
          }
        },
        "required": [
          "artist",
          "releases",
          "contracts"
        ],
        "title": "ArtistDetailOut",
        "type": "object"
      },
      "ArtistListOut": {
        "properties": {
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistOut"
            },
            "title": "Artists",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "artists",
          "total"
        ],
        "title": "ArtistListOut",
        "type": "object"
      },
      "ArtistOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "joined_at": {
            "format": "date",
            "title": "Joined At",
            "type": "string"
          },
          "legal_name": {
            "title": "Legal Name",
            "type": "string"
          },
          "n_contracts": {
            "default": 0,
            "title": "N Contracts",
            "type": "integer"
          },
          "n_releases": {
            "default": 0,
            "title": "N Releases",
            "type": "integer"
          },
          "n_tracks": {
            "default": 0,
            "title": "N Tracks",
            "type": "integer"
          },
          "stage_name": {
            "title": "Stage Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stage_name",
          "legal_name",
          "joined_at"
        ],
        "title": "ArtistOut",
        "type": "object"
      },
...
      "ClauseOut": {
        "description": "One contract clause, resolved from a structural citation (FBR-C-00501 §3).",
        "properties": {
          "artist_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Id"
          },
          "clause_no": {
            "title": "Clause No",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "contract_id": {
            "title": "Contract Id",
            "type": "integer"
          },
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "heading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "stage_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Name"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "code",
          "contract_id",
          "clause_no",
          "heading",
          "text",
          "artist_id",
          "stage_name",
          "kind",
          "effective_from",
          "effective_to"
        ],
        "title": "ClauseOut",
        "type": "object"
      },
      "ContractOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "doc_path": {
            "title": "Doc Path",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "kind",
          "effective_from",
          "effective_to",
          "doc_path"
        ],
        "title": "ContractOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReleaseListOut": {
        "properties": {
          "releases": {
            "items": {
              "$ref": "#/components/schemas/ReleaseOut"
            },
            "title": "Releases",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "releases",
          "total"
        ],
        "title": "ReleaseListOut",
        "type": "object"
      },
      "ReleaseOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "imprint": {
            "title": "Imprint",
            "type": "string"
          },
          "n_tracks": {
            "default": 0,
            "title": "N Tracks",
            "type": "integer"
          },
          "release_date": {
            "format": "date",
            "title": "Release Date",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "upc": {
            "title": "Upc",
            "type": "string"
          }
        },
        "required": [
          "id",
          "upc",
          "title",
          "imprint",
          "release_date"
        ],
        "title": "ReleaseOut",
        "type": "object"
      },
...
      "TrackListOut": {
        "properties": {
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "tracks",
          "total"
        ],
        "title": "TrackListOut",
        "type": "object"
      },
      "TrackOut": {
        "properties": {
          "duration_s": {
            "title": "Duration S",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "isrc": {
            "title": "Isrc",
            "type": "string"
          },
          "primary_artist_id": {
            "title": "Primary Artist Id",
            "type": "integer"
          },
          "stage_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Name"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isrc",
          "title",
          "duration_s",
          "primary_artist_id"
        ],
        "title": "TrackOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agent platform for music label operations — contracts, catalog, royalty statements. Agents propose; humans approve. Everything is traced.",
    "title": "Backline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/catalog/artists": {
      "get": {
        "operationId": "list_artists_catalog_artists_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artists",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/artists/{artist_id}": {
      "get": {
        "operationId": "get_artist_catalog_artists__artist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artist_id",
            "required": true,
            "schema": {
              "title": "Artist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artist",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/clauses/{code}/{clause_no}": {
      "get": {
        "description": "Resolve a structural citation to its clause text (chunks are the source —\nthe same text retrieval served the agent).",
        "operationId": "get_clause_catalog_clauses__code___clause_no__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "clause_no",
            "required": true,
            "schema": {
              "title": "Clause No",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClauseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Clause",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/releases": {
      "get": {
        "operationId": "list_releases_catalog_releases_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "imprint",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Imprint"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Releases",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/tracks": {
      "get": {
        "operationId": "list_tracks_catalog_tracks_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "artist_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artist Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tracks",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}